Use multi stage build with venv and non-root user
The goal with multi-stage is to reduce the end image size.  With the
current packages (torch + jupyter), the single stage image was 8GB.

I tried using wheels in stage 1 and installing from the wheels in the
final stage.  That reduced the size to 5.9 GB.  Using a venv came in
smaller at ~5.3 GB.

This creates a root venv at /opt/venv, and the a user venv at /app/venv.
This allows new modules to be added via jupyter, with those going into
the user venv folder which is intended to be a host volume.  Note that
I needed to install jupyter as the user and into /app/venv for this to
work properly, so the jupyter install has to be part of the final stage.